Pathfinding stuff.
Former-commit-id: a63f4be5d0551401743cf99520f6d9ac80227a99
Former-commit-id: 9a86cf2902e317beae0940062c51565bffc123a8 [formerly f932f53cbe8ab3bb908c81a0bd831bbb35049aeb]
Former-commit-id: 6fab289725ccc676ae5905eaa842064cf4c34450

+++ Source/Utility/generalUtility.h
@@ -155,7 +155,7 @@ namespace MV {
 		Type deltaX = x1-x2, deltaY = y1-y2;
 		if(deltaX<0){deltaX*=-1;}
 		if(deltaY<0){deltaY*=-1;}
-		return sqrt(((deltaX)*(deltaX)) + ((deltaY)*(deltaY)));
+		return sqrt(static_cast<double>(((deltaX)*(deltaX)) + ((deltaY)*(deltaY))));
 	}
 
 	template <class Type>
